+++ test/widget_test.dart
@@ -2,8 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:example/src/view.dart';
-
+/// Calls the example app's very own test
 import '../example/test/widget_test.dart';
 
-void main() => testApp(MyApp(key: UniqueKey()));
+void main() => testMyApp();