+++ example/CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 1.2.0
+ January 17, 2022
+- Introduced test_listener.dart
+- isInstanceOf is isA
+- import 'package:integration_test/integration_test.dart';
+
 ## 1.1.0
  January 16, 2022
 - Use tester.firstState to retrieve 'App State object'